+++ app/upgrades/v6_4/upgrade.go
@@ -1,4 +1,4 @@
-package v8
+package v6_4
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
